expansion extensions and merge library
