--- Ivyl/IOnGetStatCoefficientsReciever.cs
@@ -1,27 +1,0 @@
-﻿using BepInEx;
-using System;
-using System.IO;
-using System.Reflection;
-using Mono.Cecil;
-using MonoMod.Cil;
-using BepInEx.Bootstrap;
-using Mono.Cecil.Cil;
-using UnityEngine;
-using System.Collections.Generic;
-using RoR2;
-using R2API.ScriptableObjects;
-using UnityEngine.AddressableAssets;
-using System.Linq;
-using R2API;
-using System.Runtime.CompilerServices;
-using System.Collections;
-using RoR2.ContentManagement;
-using UnityEngine.Events;
-
-namespace Ivyl
-{
-    public interface IOnGetStatCoefficientsReciever
-    {
-        void OnGetStatCoefficients(RecalculateStatsAPI.StatHookEventArgs args);
-    }
-}
