+++ Ivyl/Ivyl.cs
@@ -1,5 +1,9 @@
 ﻿using BepInEx;
 using System;
+using RoR2;
+using System.Security.Permissions;
+using System.Security;
+using UnityEngine.ResourceManagement;
 using UnityEngine;
 using RoR2.ContentManagement;
 using HG;
@@ -14,13 +18,98 @@ using System.Text.RegularExpressions;
 using BepInEx.Bootstrap;
 using HG.GeneralSerializer;
 using UnityEngine.Rendering;
-using RoR2;
+using UnityEngine.Networking;
 using System.Threading.Tasks;
+
+[module: UnverifiableCode]
+#pragma warning disable
+[assembly: SecurityPermission(System.Security.Permissions.SecurityAction.RequestMinimum, SkipVerification = true)]
+#pragma warning restore
+[assembly: HG.Reflection.SearchableAttribute.OptIn]
 
 namespace Ivyl
 {
-    public static class MiscellaneousExtensions
+    public static class Ivyl
     {
+        public static ModelPanelParameters SetupModelPanelParameters(GameObject model, Vector3 modelRotation, float minDistance, float maxDistance, Transform focusPoint = null, Transform cameraPosition = null)
+        {
+            return SetupModelPanelParameters(model, Quaternion.Euler(modelRotation), minDistance, maxDistance, focusPoint, cameraPosition);
+        }
+
+        public static ModelPanelParameters SetupModelPanelParameters(GameObject model, ModelPanelParams info)
+        {
+            return SetupModelPanelParameters(model, info.modelRotation, info.minDistance, info.maxDistance, info.focusPoint, info.cameraPosition);
+        }
+
+        public static ModelPanelParameters SetupModelPanelParameters(GameObject model, Quaternion modelRotation, float minDistance, float maxDistance, Transform focusPoint = null, Transform cameraPosition = null)
+        {
+            ModelPanelParameters parameters = model.AddComponent<ModelPanelParameters>();
+            parameters.modelRotation = modelRotation;
+            parameters.minDistance = minDistance;
+            parameters.maxDistance = maxDistance;
+            parameters.focusPointTransform = focusPoint ?? model.transform.Find("FocusPoint") ?? model.transform.Find("Focus Point");
+            if (!parameters.focusPointTransform)
+            {
+                Transform newFocusPoint = new GameObject("FocusPoint").transform;
+                newFocusPoint.SetParent(model.transform);
+                parameters.focusPointTransform = newFocusPoint;
+            }
+            parameters.cameraPositionTransform = cameraPosition ?? model.transform.Find("CameraPosition") ?? model.transform.Find("Camera Position");
+            if (!parameters.cameraPositionTransform)
+            {
+                Transform newCameraPosition = new GameObject("CameraPosition").transform;
+                newCameraPosition.SetParent(model.transform);
+                newCameraPosition.localPosition = model.transform.forward;
+                parameters.cameraPositionTransform = newCameraPosition;
+            }
+            return parameters;
+        }
+
+        public static ItemDisplay SetupItemDisplay(GameObject displayModelPrefab)
+        {
+            ItemDisplay itemDisplay = displayModelPrefab.AddComponent<ItemDisplay>();
+            itemDisplay.rendererInfos = displayModelPrefab.GetComponentsInChildren<Renderer>()
+                .Where(x => x is MeshRenderer or SkinnedMeshRenderer)
+                .Select(x => new CharacterModel.RendererInfo
+                {
+                    defaultMaterial = x.sharedMaterial,
+                    renderer = x,
+                    defaultShadowCastingMode = ShadowCastingMode.On,
+                    ignoreOverlays = false
+                })
+                .ToArray();
+            return itemDisplay;
+        }
+
+        public static SkillFamily FindSkillFamily(GameObject bodyPrefab, SkillSlot slot)
+        {
+            if (bodyPrefab.TryGetComponent(out SkillLocator skillLocator))
+            {
+                return skillLocator.GetSkill(slot)?.skillFamily;
+            }
+            return null;
+        }
+
+        public static float StackScaling(float baseValue, float stackValue, int stack)
+        {
+            if (stack > 0)
+            {
+                return baseValue + ((stack - 1) * stackValue);
+            }
+            return 0f;
+        }
+
+        public static int StackScaling(int baseValue, int stackValue, int stack)
+        {
+            if (stack > 0)
+            {
+                return baseValue + ((stack - 1) * stackValue);
+            }
+            return 0;
+        }
+
+        public static bool IsModLoaded(string guid) => Chainloader.PluginInfos.ContainsKey(guid);
+
         public static void Add<TAsset>(this NamedAssetCollection<TAsset> assetCollection, TAsset newAsset)
         {
             string assetName = assetCollection.nameProvider(newAsset);
@@ -78,7 +167,7 @@ namespace Ivyl
 
         public static int AddScene(this SceneCollection sceneCollection, SceneDef scene, float weight = 1f)
         {
-            ArrayUtils.ArrayAppend(ref sceneCollection._sceneEntries, new SceneCollection.SceneEntry 
+            ArrayUtils.ArrayAppend(ref sceneCollection._sceneEntries, new SceneCollection.SceneEntry
             {
                 sceneDef = scene,
                 weight = weight,
@@ -88,10 +177,10 @@ namespace Ivyl
 
         public static int AddRelationshipPair(this ItemRelationshipProvider itemRelationshipProvider, ItemDef item1, ItemDef item2)
         {
-            return AddRelationshipPair(itemRelationshipProvider, new ItemDef.Pair 
-            { 
-                itemDef1 = item1, 
-                itemDef2 = item2 
+            return AddRelationshipPair(itemRelationshipProvider, new ItemDef.Pair
+            {
+                itemDef1 = item1,
+                itemDef2 = item2
             });
         }
 
@@ -199,7 +288,7 @@ namespace Ivyl
         }
 
         public static bool HasItem(this CharacterBody characterBody, ItemDef itemDef, out int stack) => HasItem(characterBody, itemDef ? itemDef.itemIndex : ItemIndex.None, out stack);
-        
+
         public static bool HasItem(this CharacterBody characterBody, ItemIndex itemIndex, out int stack)
         {
             if (characterBody && characterBody.inventory)
@@ -216,7 +305,7 @@ namespace Ivyl
         public static bool HasItem(this CharacterBody characterBody, ItemIndex itemIndex) => characterBody && characterBody.inventory && characterBody.inventory.GetItemCount(itemIndex) > 0;
 
         public static bool HasItem(this CharacterMaster characterMaster, ItemDef itemDef, out int stack) => HasItem(characterMaster, itemDef ? itemDef.itemIndex : ItemIndex.None, out stack);
-        
+
         public static bool HasItem(this CharacterMaster characterMaster, ItemIndex itemIndex, out int stack)
         {
             if (characterMaster && characterMaster.inventory)
@@ -242,9 +331,9 @@ namespace Ivyl
         }
 
         public static bool HasItem(this CharacterMaster characterMaster, ItemDef itemDef) => HasItem(characterMaster, itemDef.itemIndex);
-        
+
         public static bool HasItem(this CharacterMaster characterMaster, ItemIndex itemIndex) => characterMaster && characterMaster.inventory && characterMaster.inventory.GetItemCount(itemIndex) > 0;
-        
+
         public static void ClearDotStacksForType(this DotController dotController, DotController.DotIndex dotIndex)
         {
             for (int i = dotController.dotStackList.Count - 1; i >= 0; i--)
@@ -267,7 +356,7 @@ namespace Ivyl
         public static void RemoveIncomingDamageReceiver(this HealthComponent healthComponent, IOnIncomingDamageServerReceiver onIncomingDamageReceiver)
         {
             if (healthComponent && Array.IndexOf(healthComponent.onIncomingDamageReceivers, onIncomingDamageReceiver) is var index && index >= 0)
-            {   
+            {
                 ArrayUtils.ArrayRemoveAtAndResize(ref healthComponent.onIncomingDamageReceivers, index);
             }
         }
